+++ calculator/public/index.html
@@ -7,7 +7,7 @@
     <title>Calculator</title>
 
     <link rel="stylesheet" href="../src/css/index.css">
-    <script src="../src/index.js" defer></script>
+    <script type="module" src="../src/index.js" defer></script>
     
     <!-- import jQuery library -->
     <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.6.4/jquery.min.js" crossorigin></script>
@@ -23,7 +23,7 @@
         </section>
 
         <section class="calc__buttons">
-          <button class="calc__copy-result" onclick="copyResultToClipboard()">copy</button>
+          <button class="calc__copy-result">copy</button>
           
           <button id="1">1</button>
           <button id="2">2</button>
