+++ calculator/src/tools/rules.js
@@ -1,0 +1,32 @@
+const LEGAL_KEYS = [
+  "0",
+  "1",
+  "2",
+  "3",
+  "4",
+  "5",
+  "6",
+  "7",
+  "8",
+  "9",
+  ".",
+  "+",
+  "-",
+  "*",
+  "/",
+  "ArrowUp",
+  "ArrowDown",
+  "ArrowLeft",
+  "ArrowRight",
+  "Home",
+  "End",
+  "Backspace",
+  "Delete",
+];
+// ["x", "X", "a", "A", "^", "!", "(", ")"];
+
+
+const OPERATORS = ["+", "-", "*", "/"];
+
+
+export { LEGAL_KEYS, OPERATORS };
